Move indy wallet related structs to wallet/indy_wallet.rs
