+++ aries_vcx/tests/test_agency.rs
@@ -17,6 +17,7 @@ mod integration_tests {
     use aries_vcx::global::settings;
     use aries_vcx::utils::devsetup::SetupPoolDirectory;
     use aries_vcx_core::wallet::agency_client_wallet::ToBaseAgencyClientWallet;
+    use aries_vcx_core::wallet::indy_wallet::WalletConfig;
     use messages::msg_fields::protocols::cred_issuance::offer_credential::OfferCredentialDecorators;
     use messages::msg_fields::protocols::cred_issuance::{CredentialAttr, CredentialPreview};
     use messages::msg_fields::protocols::notification::Notification;
@@ -450,7 +451,7 @@ mod integration_tests {
     #[ignore]
     async fn test_agency_pool_update_agent_webhook() {
         use aries_vcx_core::{
-            indy::wallet::{create_and_open_wallet, WalletConfig},
+            indy::wallet::{create_and_open_wallet},
             wallet::{base_wallet::BaseWallet, indy_wallet::IndySdkWallet},
         };
