+++ aries_vcx/tests/test_mysql_wallet.rs
@@ -13,10 +13,10 @@ mod dbtests {
     use aries_vcx::utils::provision::provision_cloud_agent;
     use aries_vcx::utils::test_logger::LibvcxDefaultLogger;
     use aries_vcx_core::indy::wallet::{
-        close_wallet, create_wallet_with_master_secret, open_wallet, wallet_configure_issuer, WalletConfig,
+        close_wallet, create_wallet_with_master_secret, open_wallet, wallet_configure_issuer,
         WalletConfigBuilder,
     };
-    use aries_vcx_core::wallet::indy_wallet::IndySdkWallet;
+    use aries_vcx_core::wallet::indy_wallet::{IndySdkWallet, WalletConfig};
 
     #[tokio::test]
     #[ignore]
