+++ aries_vcx/tests/utils/devsetup_faber.rs
@@ -25,14 +25,15 @@ use aries_vcx::protocols::mediated_connection::inviter::state_machine::InviterSt
 use aries_vcx::protocols::proof_presentation::verifier::state_machine::VerifierState;
 use aries_vcx::protocols::proof_presentation::verifier::verification_status::PresentationVerificationStatus;
 use aries_vcx::protocols::revocation_notification::sender::state_machine::SenderConfigBuilder;
-use aries_vcx::utils::devsetup::{SetupPoolDirectory, SetupProfile, AGENCY_DID, AGENCY_ENDPOINT, AGENCY_VERKEY};
+use aries_vcx::utils::devsetup::{AGENCY_DID, AGENCY_ENDPOINT, AGENCY_VERKEY, SetupPoolDirectory, SetupProfile};
 use aries_vcx::utils::provision::provision_cloud_agent;
 use aries_vcx_core::indy::wallet::{
-    create_wallet_with_master_secret, open_wallet, wallet_configure_issuer, IssuerConfig, WalletConfig,
+    create_wallet_with_master_secret, open_wallet, wallet_configure_issuer,
 };
 #[cfg(feature = "modular_libs")]
 use aries_vcx_core::ledger::request_submitter::vdr_ledger::LedgerPoolConfig;
 use aries_vcx_core::{PoolHandle, WalletHandle};
+use aries_vcx_core::wallet::indy_wallet::{IssuerConfig, WalletConfig};
 use diddoc_legacy::aries::service::AriesService;
 use messages::decorators::please_ack::AckOn;
 use messages::msg_fields::protocols::connection::invitation::public::{PublicInvitation, PublicInvitationContent};
