+++ libvcx/src/api_c/wallet.rs
@@ -1,10 +1,10 @@
 use std::ptr::null;
 
-use aries_vcx::aries_vcx_core::{SearchHandle, WalletHandle, INVALID_SEARCH_HANDLE};
+use aries_vcx::aries_vcx_core::{INVALID_SEARCH_HANDLE, SearchHandle, WalletHandle};
 use futures::future::BoxFuture;
 use libc::c_char;
 
-use aries_vcx::aries_vcx_core::indy::wallet::{RestoreWalletConfigs, WalletConfig};
+use aries_vcx::aries_vcx_core::wallet::indy_wallet::{RestoreWalletConfigs, WalletConfig};
 
 use libvcx_core;
 use libvcx_core::errors::error::{LibvcxError, LibvcxErrorKind};
