+++ libvcx_core/src/api_vcx/api_global/state.rs
@@ -5,10 +5,11 @@ use crate::api_vcx::api_global::settings::get_config_value;
 use crate::api_vcx::api_global::wallet::close_main_wallet;
 use crate::errors::error::LibvcxResult;
 use aries_vcx::aries_vcx_core::indy::ledger::pool;
-use aries_vcx::aries_vcx_core::indy::wallet::{delete_wallet, WalletConfig};
+use aries_vcx::aries_vcx_core::indy::wallet::delete_wallet;
+use aries_vcx::aries_vcx_core::wallet::indy_wallet::WalletConfig;
 use aries_vcx::global::settings::{
-    reset_config_values_ariesvcx, CONFIG_WALLET_KEY, CONFIG_WALLET_KEY_DERIVATION, CONFIG_WALLET_NAME,
-    CONFIG_WALLET_TYPE, DEFAULT_POOL_NAME, DEFAULT_WALLET_NAME, UNINITIALIZED_WALLET_KEY, WALLET_KDF_DEFAULT,
+    CONFIG_WALLET_KEY, CONFIG_WALLET_KEY_DERIVATION, CONFIG_WALLET_NAME, CONFIG_WALLET_TYPE,
+    DEFAULT_POOL_NAME, DEFAULT_WALLET_NAME, reset_config_values_ariesvcx, UNINITIALIZED_WALLET_KEY, WALLET_KDF_DEFAULT,
 };
 
 pub fn state_vcx_shutdown() {
