+++ wrappers/vcx-napi-rs/src/api/wallet.rs
@@ -1,11 +1,11 @@
 use libvcx_core::api_vcx::api_global::settings::settings_init_issuer_config;
 use libvcx_core::api_vcx::api_global::{ledger, wallet};
-use libvcx_core::aries_vcx::aries_vcx_core::indy::wallet::{IssuerConfig, RestoreWalletConfigs, WalletConfig};
 use libvcx_core::errors::error::{LibvcxError, LibvcxErrorKind};
 use libvcx_core::serde_json;
 use libvcx_core::serde_json::json;
 use napi::bindgen_prelude::Buffer;
 use napi_derive::napi;
+use libvcx_core::aries_vcx::aries_vcx_core::wallet::indy_wallet::{IssuerConfig, RestoreWalletConfigs, WalletConfig};
 
 use crate::error::to_napi_err;
